Tk002

+++ src/main/java/com/spring/jwt/config/filter/JwtRefreshTokenFilter.java
@@ -231,7 +231,7 @@ public class JwtRefreshTokenFilter extends AbstractAuthenticationProcessingFilte
             accessTokenCookie.setHttpOnly(false);
             accessTokenCookie.setSecure(true);
             accessTokenCookie.setPath("/");
-            accessTokenCookie.setMaxAge(jwtConfig.getExpiration());
+            accessTokenCookie.setMaxAge((int) jwtConfig.getExpiration());
             response.addCookie(accessTokenCookie);
 
             Map<String, Object> tokens = new HashMap<>();
